Preserve storage offset on meta conversions.
This PR modifies the `to_meta` function present in the functionalization layer, so that it
also preserves the storage offset.

It does so by creating a base tensor, with enough elements for a view with the source
tensor's metadata to live. Then, we follow that by calling `as_strided`, resulting in a
tensor with the expected metadata.

ghstack-source-id: 4f21efca7ba8d96dc71b0e7a5969764bf1b95478
Pull Request resolved: https://github.com/pytorch/pytorch/pull/135498